+++ grammar/ChangeLog.md
@@ -1,5 +1,9 @@
 This file logs the changes in FST files, starting from the DSDSmor fork.
 
+2022-09-20:
+
+* replace `<Pred>` and `<Adv>` with `<NonAttr>`
+
 2022-09-09:
 
 * allow for genitive e-elision after diphthongs "ai" and "ei"